pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crisis-inbox"
version = "0.1.0"
description = "RL environment for training LLMs to manage personal task overload during natural disasters"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "crisis_inbox.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["crisis_inbox", "crisis_inbox.server"]
package-dir = { "crisis_inbox" = ".", "crisis_inbox.server" = "server" }
